Move ShuttingDownException to common and fix a broken test
