+++ core/src/main/java/com/linecorp/armeria/server/GracefulShutdown.java
@@ -21,6 +21,7 @@ import java.time.Duration;
 import com.linecorp.armeria.common.HttpRequest;
 import com.linecorp.armeria.common.HttpResponse;
 import com.linecorp.armeria.common.HttpStatus;
+import com.linecorp.armeria.common.ShuttingDownException;
 import com.linecorp.armeria.common.annotation.UnstableApi;
 
 /**
