+++ core/src/main/java/com/linecorp/armeria/server/GracefulShutdownBuilder.java
@@ -25,6 +25,7 @@ import java.util.function.BiFunction;
 
 import com.linecorp.armeria.common.HttpRequest;
 import com.linecorp.armeria.common.HttpStatus;
+import com.linecorp.armeria.common.ShuttingDownException;
 import com.linecorp.armeria.common.annotation.UnstableApi;
 
 /**
