+++ core/src/test/java/com/linecorp/armeria/GracefulShutdownBuilderTest.java
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Test;
 
 import com.linecorp.armeria.internal.testing.AnticipatedException;
 import com.linecorp.armeria.server.GracefulShutdown;
-import com.linecorp.armeria.server.ShuttingDownException;
+import com.linecorp.armeria.common.ShuttingDownException;
 
 class GracefulShutdownBuilderTest {
